doc-processor/requirements.txt
google-cloud-documentai
google-cloud-storage
google-cloud-bigquery